% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.stats.R
\name{compare.stats}
\alias{compare.stats}
\title{Group and compare summaries statistics to a data frame}
\usage{
compare.stats(df, group_var, values, m_functions)
}
\arguments{
\item{df}{numeric vector or matrix or dataframe}

\item{group_var}{character vector of variables to group the data}

\item{values}{numeric or integer variables}

\item{m_functions}{functions to apply in values}
}
\value{
This function returns a data frame with the selected group_vars and 
the result of m_functions applied in the selected values.
}
\description{
Extract and return a data frame with the columns that include only numeric values
}
\author{
Kleanthis Koupidis
}
