% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Coefficient of variation}
\usage{
CV(x)
}
\arguments{
\item{x}{A numeric vector or matrix or dataframe}
}
\value{
This function returns a vector with the coefficient of variance for the input vector,matrix or data frame.
}
\description{
Extract and return a data frame with the columns that include only numeric values
}
\author{
Kleanthis Koupidis
}
