% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.analysis.R
\name{ds.analysis}
\alias{ds.analysis}
\title{Calculation of some Descriptive Tasks}
\usage{
ds.analysis(data, c.out=1.5, box.width=0.15, outliers=T,
 corr.method= "pearson", fr.select=NULL)
}
\arguments{
\item{data}{The input data}

\item{c.out}{Determines the length of the "whiskers" plot.
If it is equal to zero no outliers will be returned.}

\item{box.width}{The width level is determined 0.15 times the square root of the size of the input data.}

\item{outliers}{If TRUE the outliers will be computed at the selected "c.out" level (default is 1.5 times the Interquartile Range).}

\item{corr.method}{The correlation coefficient method to compute: "pearson" (default),
"kendall" or "spearman".}

\item{fr.select}{One or more nominal variables to calculate their corresponding frequencies.}
}
\value{
A list with...
}
\description{
The input of this function can be a matrix or data frame. The function calculates the basic descriptive measures, 
the correlation and the boxplot parameters of all the numerical variables, the frequencies of all the nominal variables
and the necessary components of the selected generilized linear model.
}
\details{
...
}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{open_spending.ds}}
}

