% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boxplot.R
\name{ds.boxplot}
\alias{ds.boxplot}
\title{Calculation of Statistics for the Boxplot}
\usage{
ds.boxplot(data, out.level=1.5)
}
\arguments{
\item{data}{The input matrix or data frame}

\item{out.level}{Determines the times of length of the box the "whiskers" are away from the box. If it is equal to zero, the whiskers extend to the data extremes (and no outliers be returned)}
}
\value{
Returns a list with the following components:

stats: a vector with the extreme of the lower whisker, the lower "hinge", the median, the upper "hinge" and the extreme of the upper whisker.

n: the number of non-NA observations in the sample.

out: the values of any data points which lie beyond the whiskers' extremes
}
\description{
This function calculates the statistics for the boxplot of the input dataset.
}
\details{
This function returns the necessary statistics for bulding a boxplot
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
\seealso{
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}

