% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.frequencies.R, R/ds.glm.R
\name{ds.frequencies}
\alias{ds.frequencies}
\alias{ds.glm}
\title{frequencies}
\usage{
ds.frequencies(data,select=NULL)

ds.glm(x, dependent=NULL, independent=NULL, distr.family = "gaussian")
}
\arguments{
\item{data}{A matrix or data frame which includes at least one factor/character.}

\item{select}{...}

\item{x}{The input matrix or data frame}

\item{dependent}{The dependent variables of the model}

\item{independent}{The independent variables of the model}

\item{distr.family}{A character string naming a the error distribution 
and link function to be used in the model(See family for details of family functions.)}
}
\description{
This functions calculates frequencies of factors/characters of the input dataset.

\code{\link{ds.glm}} is used to fit generalized linear models through glm from stats package and 
return some results in json format.
}
\details{
This function returns a list with the frequencies of factors/characters of the input dataset.

This function returns a list with the frequencies of factors/characters of the input dataset.
}
\author{
Kleanthis Koupidis

Kleanthis Koupidis
}

