% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.glm.R
\name{ds.glm}
\alias{ds.glm}
\title{Generalized Linear Models}
\usage{
ds.glm(x, dependent=NULL, independent=NULL, distr.family = "gaussian")
}
\arguments{
\item{x}{The input matrix or data frame}

\item{dependent}{The dependent variables of the model}

\item{independent}{The independent variables of the model}

\item{distr.family}{A character string naming a the error distribution 
and link function to be used in the model(See family for details of family functions.)}
}
\value{
coefficients
 residuals
 fitted
 residuals.degfred
 qq.plot
 linear.predictors
}
\description{
\code{\link{ds.glm}} is used to fit generalized linear models through glm from stats package and 
return some results in json format.
}
\details{
Generalized linear models are used to fit data, using glm stats package, by specifying the
dependent and independent variables and the description of the distribution error 
(default is gaussian). If user provides only the dependent variables, the rest 
variables are selected as independent. If user provides the independent variables, 
the rest variables are selected as dependent. If user provides the independent variables, 
the two variables with the highest correlation are selected.
}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}

