% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.hist.R
\name{ds.hist}
\alias{ds.hist}
\title{Histogram breaks and frequencies}
\usage{
ds.hist(x, breaks="Sturges")
}
\arguments{
\item{x}{The input numeric vector.}

\item{breaks}{The method or the number of classes for the histogram.}
}
\value{
A list with the following components:

\itemize{
\item cuts The boundaries of the histogram classes
\item counts The frequency of each histogram class
\item normal.curve The normal curve 
\item mean The average value of the input vector
\item median The median value of the input data
}
}
\description{
This function computes the histogram parameters of the numeric input vector. The default for 
breaks is the value resulted from "Sturges" algorithm.
}
\details{
The possible values for breaks are "Sturges"(see nclass.Sturges), "Scott"(see nclass.scott),
"FD" / "Freedman-Diaconis" (see nclass.FD) and other integer value.
}
\author{
Kleanthis Koupidis
}
\seealso{
\code{\link{ds.analysis}},\code{\link{open_spending.ds}}
}

