% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.skewness.R
\name{ds.skewness}
\alias{ds.skewness}
\title{Calculation of Skewness}
\usage{
ds.skewness(x)
}
\arguments{
\item{x}{A vector, matrix or data frame.}
}
\description{
This function calculates skewness of the input vector, matrix or data frame.
}
\details{
This function returns the skewness, also known as Pearson's moment coefficient of skewness, of numbers of the input data.
}
\author{
Aikaterini Chatzopoulou
}
\seealso{
\code{\link{ds.kurtosis}}, \code{\link{ds.statistics}}, 
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}

