% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisub.R
\name{multisub}
\alias{multisub}
\title{Multiple replacement}
\usage{
multisub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{Chararcter string vector containing a regular expression to be matched in the given character vector}

\item{replacement}{A character vector of equal length with the pattern to be replaced.}

\item{x}{A character vector or an object where the matches are}

\item{...}{other parameters to pass}
}
\value{
This function returns a character vector with the replacements.
}
\description{
Extract and return a data frame with the columns that include only numeric values
}
\author{
Kleanthis Koupidis
}
