% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{non_nums}
\alias{non_nums}
\title{Select the non numeric columns of a given dataset}
\usage{
non_nums(data)
}
\arguments{
\item{data}{A vector, matrix or data frame.}
}
\value{
This function returns a data frame with the non numeric columns of the input dataset.
}
\description{
Extract and return a data frame with the columns that include non numeric values
}
\examples{
# with data frame as input
non_nums(iris)


# OpenBudgets.eu Dataset Example:
head(nums(Wuppertal_df))
}
\author{
Kleanthis Koupidis
}
