% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nums.R
\name{nums}
\alias{nums}
\title{Select the numeric columns of a given dataset}
\usage{
nums(data)
}
\arguments{
\item{data}{The json string, URL or file from Open Spending API}
}
\value{
This function returns a dataframe with the numeric columns of the input dataset.
}
\description{
Extract and return a dataframe with the columns that include only numeric values
}
\author{
Kleanthis Koupidis
}

