% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_df_openspending.R
\name{sample_df_openspending}
\alias{sample_df_openspending}
\title{Sample data from OpenSpending}
\format{A link to json format file}
\source{
Open Spending
}
\description{
Sample data of Revised Budget phase amounts 
\itemize{
  \item The functional classification function
  \item The original amounts of this year
  \item The functional classification code
   \item The year (2016) of the recorded approved budget phase amounts
  \item The revised budget phase amounts of 2016
  }
}

