% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_json_openspending.R
\name{sample_json_openspending}
\alias{sample_json_openspending}
\title{Sample data from Open Spending}
\format{A json format file}
\source{
OpenSpending
}
\description{
Sample data of Revised Budget phase amounts
\itemize{
  \item The year (2016) of the recorded approved budget phase amounts
  \item The revised budget phase amounts of 2016
  \item The original amounts of this year
  \item The functional classification description
  \item The functional classification code
  }
}

