% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wuppertal_df.R
\name{Wuppertal_df}
\alias{Wuppertal_df}
\title{Wuppertal Fiscal Data extracted from Open Spending API}
\format{A link with the json format data}
\source{
OpenSpending
}
\description{
...
\itemize{
  \item The year ...
  \item The ....
  \item The ...
  \item The ...
  \item The ...
  }
}
