% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wuppertal_openspending.R
\name{Wuppertal_openspending}
\alias{Wuppertal_openspending}
\title{Wuppertal Fiscal Data extracted from Open Spending API}
\format{A link with the json format data}
\source{
OpenSpending
}
\description{
This dataset contains the the budget for wuppertal for 2009 to 2018, as extracted in 2014-2015.

\itemize{
  \item The product ID
  \item The account type
  \item The kind
  \item The year these amounts were measaured
  \item The amount
  \item The product area ID
  \item The product group ID
  \item The product
  \item The product area
  \item The product group
  }
}
