% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boxplot.R
\name{ds.boxplot}
\alias{ds.boxplot}
\title{Boxplot}
\usage{
ds.boxplot( data, q = 1.5 )
}
\arguments{
\item{data}{The input matrix or data frame}

\item{q}{The dependent variables of the model}
}
\description{
Boxplot
}
\details{
boxplot
}
\author{
Kleanthis Koupidis
}

