% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.boxplot.R
\name{ds.boxplot}
\alias{ds.boxplot}
\title{Boxplot}
\usage{
ds.boxplot(data, q=1.5)
}
\arguments{
\item{data}{The input matrix or data frame}

\item{q}{...}
}
\value{
Description of the returns
}
\description{
Boxplot
}
\details{
boxplot
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
\seealso{
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}

