% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.correlation.R
\name{ds.correlation}
\alias{ds.correlation}
\title{Correlation Coefficient of a dataframe}
\usage{
ds.correlation(x, y=NULL, cor.method="pearson", tojson=F)
}
\arguments{
\item{x}{A vector, matrix or data frame}

\item{y}{A vector, matrix or data frame}

\item{cor.method}{The correlation coefficient method to compute: "pearson" (default),
"kendall" or "spearman".}

\item{tojson}{If TRUE the results are returned in json format, default returns a data frame}
}
\description{
This functions calculates the correlation coefficient of the input vectors, matrix or data frame.
By default, the correlation coefficient of pearson is computed.
}
\details{
This function returns an upper triangle matrix with the correlation coefficients 
of the input data. The correlation coefficient of pearson is computed, by default. 
Other options are "kendall" or "spearman".
}
\seealso{
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
