% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.correlation.R
\name{ds.correlation}
\alias{ds.correlation}
\title{Correlation Coefficient}
\usage{
ds.correlation(x, y=NULL, cor.method="pearson")
}
\arguments{
\item{x}{A vector, matrix or data frame}
}
\description{
This functions calculates the correlation coefficient of the input vectors or datasets.
By default, the correlation coefficient of pearson is computed.
}
\details{
This function returns a symmetric matrix in json format with the correlation coefficients 
of the input data. The correlation coefficient of pearson is computed, by default. 
Other options are "kendall" or "spearman".
}
\author{
Aikaterini Chatzopoulou, Kleanthis Koupidis
}
\seealso{
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}

