% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.glm.R
\name{ds.glm}
\alias{ds.glm}
\title{Generalized Linear Models}
\usage{
ds.glm(x, dependent=NULL, independent=NULL, distr.family = "gaussian")
}
\arguments{
\item{x}{The input matrix or data frame}

\item{dependent}{The dependent variables of the model}

\item{independent}{The independent variables of the model}

\item{distr.family}{A character string naming a the error distribution 
and link function to be used in the model(See family for details of family functions.)}
}
\description{
\code{\link{ds.glm}} is used to fit generalized linear models through glm from stats package and 
return some results in json format.
}
\details{
This function returns a list with the frequencies of factors/characters of the input dataset.
}
\author{
Kleanthis Koupidis
}

