% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.frequency.R
\name{ds.frequency}
\alias{ds.frequency}
\title{Frequencies of nominal variables}
\usage{
ds.frequency(data,select=NULL)
}
\arguments{
\item{data}{A vector, matrix or data frame which includes at least one factor/character.}

\item{select}{One or more nominal variables to calculate their corresponding frequencies.}
}
\description{
This functions calculates frequencies of factors/characters of the input dataset.
}
\details{
This function returns a json output with the frequencies of factors/characters of the input dataset.
}
\author{
Kleanthis Koupidis
}

