% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.kurtosis.R
\name{ds.kurtosis}
\alias{ds.kurtosis}
\title{Kurtosis}
\usage{
ds.kurtosis(x)
}
\arguments{
\item{x}{A vector, matrix or data frame}
}
\description{
This functions calculates kurtosis of the input dataset.
}
\details{
This function returns a json output with the kurtosis of numbers of the input data
}
\author{
Aikaterini Chatzopoulou
}

