% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.kurtosis.R
\name{ds.kurtosis}
\alias{ds.kurtosis}
\title{Calculation of Kurtosis}
\usage{
ds.kurtosis(x)
}
\arguments{
\item{x}{A vector, matrix or data frame.}
}
\description{
This function calculates kurtosis of the input vector, matrix or data frame.
}
\details{
This function returns the kurtosis of numbers of the input data.
}
\author{
Aikaterini Chatzopoulou
}
\seealso{
\code{\link{ds.skewness}}, \code{\link{ds.statistics}}, 
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}

