% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.skewness.R
\name{ds.skewness}
\alias{ds.skewness}
\title{Calculation of Skewness}
\usage{
ds.skewness(x, tojson = FALSE)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{tojson}{If TRUE the results are returned in json format}
}
\description{
This function calculates skewness of the input vector, matrix or data frame.
}
\details{
This function returns the skewness, also known as Pearson's moment coefficient of skewness, of numbers of the input data.
}
\examples{
# with a matrix as an input
Matrix <- cbind(
  Uni05 = (1:200) / 21, Norm = rnorm(200),
  `5T` = rt(200, df = 5), Gam2 = rgamma(200, shape = 2)
)
ds.skewness(Matrix, tojson = FALSE)

# with iris data frame as an input
ds.skewness(iris, tojson = FALSE)

# with a vector as an input and json output
vec <- as.vector(iris$Sepal.Width)
ds.skewness(vec, tojson = TRUE)

# OpenBudgets.eu Dataset Example:
ds.skewness(Wuppertal_df, tojson = FALSE)
}
\seealso{
\code{\link{ds.kurtosis}}, \code{\link{ds.statistics}},
\code{\link{ds.analysis}}, \code{\link{open_spending.ds}}
}
\author{
Aikaterini Chatzopoulou
}
