% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.skewness.R
\name{ds.skewness}
\alias{ds.skewness}
\title{Skewness}
\usage{
ds.skewness(x)
}
\arguments{
\item{x}{A vector, matrix or data frame which includes at least one number.}
}
\description{
This functions calculates skewness of the input dataset.
}
\details{
This function returns a json output with the skewness of numbers of the input dataset.
}
\author{
Kleanthis Koupidis
}

