% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.statistics.R
\name{ds.statistics}
\alias{ds.statistics}
\title{Calculation of the Statistic Measures}
\usage{
ds.statistics(data)
}
\arguments{
\item{data}{A numeric vector, matrix or data frame}
}
\value{
A json file with the following components:
\itemize{
\item Min: The minimum observed value of the input data
\item Max: The maximum observed value of the input data
\item Range: The range, defined as the difference of the maximum and the minimum value.
\item Mean: The average value of the input data
\item Median: The median value of the input data
\item Quantiles: The 0\%,25\%,50\%,75\%,100\% percentiles
\item Variance: The variance of the input data
\item StandardDeviation: The standard deviation of the input data
\item Skewness: The Skewness of the input data
\item Kurtosis: The Kurtosis of the input data
}
}
\description{
This function calculates the basic descriptive measures of the input dataset.
}
\details{
This function returns the min, max, range, mean, median, 0\%,25\%,50\%,75\%,100\% quantiles variance,
standartdeviation, skewness and kurtosis of the input data.
}
\author{
Katerina Chatzopoulou, Kleanthis Koupidis
}
\seealso{
\code{\link{open_spending.ds}}
}

