% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.analysis.R, R/open_spending.ds.R
\name{ds.analysis}
\alias{ds.analysis}
\alias{open_spending.ds}
\title{Read and Calculate the Basic Information for Basic Descriptive Tasks from Open Spending API}
\usage{
ds.analysis(json_data,method)

open_spending.ds(json_data,amount)
}
\arguments{
\item{json_data}{The json string, URL or file from Open Spending API}

\item{amount}{Specify the amount label of the json input data}

\item{json_data}{The json string, URL or file from Open Spending API}

\item{amount}{Specify the amount label of the json input data}
}
\value{
A json string with the resulted parameters of the \code{\link{ds.analysis}} function.

A json string with the resulted parameters of the \code{\link{ds.analysis}} function.
}
\description{
Extract and analyze the input data provided from Open Spending API, using the ds.analysis function.

Extract and analyze the input data provided from Open Spending API, using the ds.analysis function.
}
\details{
This function is used to read data in json format from Open Spending API, in order to implement 
some basic descriptive tasks through \code{\link{ds.analysis}} function.

This function is used to read data in json format from Open Spending API, in order to implement 
some basic descriptive tasks through \code{\link{ds.analysis}} function.
}
\author{
Kleanthis Koupidis

Kleanthis Koupidis
}
\seealso{
\code{\link{ds.analysis}}

\code{\link{ds.analysis}}
}

