% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.analysis.R, R/open_spending.ds.R
\name{ds.analysis}
\alias{ds.analysis}
\alias{open_spending.ds}
\title{Read and Calculate the Basic Information for Basic Descriptive Tasks from Open Spending API}
\usage{
ds.analysis(data, box.out=1.5, corr.method= "pearson", fr.select=NULL)

open_spending.ds(json_data, box.outl=1.5, cor.method= "pearson", select=NULL)
}
\arguments{
\item{data}{The input data}

\item{box.out}{...}

\item{corr.method}{The correlation coefficient method to compute: "pearson" (default),
"kendall" or "spearman".}

\item{fr.select}{One or more nominal variables to calculate their corresponding frequencies.}

\item{json_data}{The json string, URL or file from Open Spending API}

\item{box.outl}{...}

\item{cor.method}{The correlation coefficient method to compute: "pearson" (default),
"kendall" or "spearman".}

\item{select}{One or more nominal variables to calculate their corresponding frequencies.}
}
\value{
A json string with the resulted parameters of the \code{\link{ds.analysis}} function.

A json string with the resulted parameters of the \code{\link{ds.analysis}} function.
}
\description{
Extract and analyze the input data provided from Open Spending API, using the ds.analysis function.

Extract and analyze the input data provided from Open Spending API, using the ds.analysis function.
}
\details{
This function is used to read data in json format from Open Spending API, in order to implement 
some basic descriptive tasks through \code{\link{ds.analysis}} function.

This function is used to read data in json format from Open Spending API, in order to implement 
some basic descriptive tasks through \code{\link{ds.analysis}} function.
}
\author{
Kleanthis Koupidis

Kleanthis Koupidis
}
\seealso{
\code{\link{open_spending.ds}}

\code{\link{ds.analysis}}
}

