% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_json_link_rudolf.r
\name{sample_json_link_rudolf}
\alias{sample_json_link_rudolf}
\title{Sample data from Rudolf API}
\format{A link with the json format data}
\source{
Rudolf
}
\description{
Sample data from Rudolf API
\itemize{
  \item The year (2016) of the recorded approved budget phase amounts
  \item The revised budget phase amounts of 2016
  \item The original amounts of this year
  \item The functional classification description
  \item The functional classification code
  }
}
