% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wuppertal_openspending.R
\name{Wuppertal_openspending}
\alias{Wuppertal_openspending}
\title{Wuppertal Fiscal Data extracted from Open Spending API}
\format{A link with the json format data}
\source{
http://next.openspending.org/api/3/cubes/4b6d969e07ef7a86aa54e539fc127a14:wuppertalhaushalt/facts
}
\description{
This dataset contains the budget of wuppertal for 2009 to 2020

\itemize{
  \item The product ID
  \item The account type
  \item The kind
  \item The year these amounts were measaured
  \item The amount
  \item The product area ID
  \item The product group ID
  \item The product
  \item The product area
  \item The product group
  }
}
